requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# AI/ML dependencies
crewai>=0.30.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.38
openai>=1.0.0
anthropic>=0.40.0
google-generativeai>=0.8.0
# Vertex AI SDK (Gemini/Imagen)
google-cloud-aiplatform>=1.64.0

# API dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
websockets>=11.0.0

# Database and storage
sqlalchemy>=2.0.0
alembic>=1.10.0
asyncpg>=0.29.0
chromadb>=0.4.0
supabase>=2.0.0

# Utilities
typer>=0.9.0
rich>=13.0.0
httpx>=0.24.0
aiofiles>=23.0.0
psutil>=5.9.0
Jinja2>=3.1.4

# Metrics and monitoring
prometheus-client>=0.20.0


requests>=2.31.0
aiohttp>=3.9.0


# Google auth for Service Account
google-auth>=2.30.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Optional: Streamlit for legacy interface
streamlit>=1.30.0
